* release/4mr.cn: (48 commits)
  替换icon
  重构场景编辑器的载入代码，改变所有处理内容皆为右手坐标系，坐标系转换工作交给unity去完成
  搞定所有的Breadcrumb
  节点相关
  更新版本
  小更新
  应该是完成了这一个版本
  用户管理界面
  明天要做user的整理
  no message
  no message
  no message
  no message
  no message
  no message
  no message
  no message
  no message
  no message
  no message
  ...

+++ src/node-editor/plugins/knight.js
@@ -17,9 +17,10 @@ function install(editor, options) {
 
     if (id > 0) {
       try {
-        alert(id)
-        await deleteMetaKnight(id)
-        console.log('delete ok' + id)
+        if (options.root.canSave) {
+          await deleteMetaKnight(id)
+          console.log('delete ok' + id)
+        }
       } catch (e) {
         console.error(e)
       }
@@ -36,12 +37,11 @@ function install(editor, options) {
     const uuid = uuidv4()
     let id = component.data['id']
 
-    if (typeof id === 'undefined') {
+    if (typeof id === 'undefined' && options.root.canSave) {
       const response = await postMetaKnight({
         verse_id: options.verseId,
         uuid
       })
-      //alert(JSON.stringify(response))
       const data = response.data
       id = data.id
       component.controls.get('id').setValue(id)
